Add Matrix class
Matrices are now available from the linalg module. Stores the elements in
a list of list of floats. Also implements a shape method which returns
a tuple of the number of rows and columns.

make_matrix: Takes in the number of rows, cols, and a function which is
used to generate the elements.

zeros: Takes in the number of rows and columns and returns a corresponding
Matrix filled with zeros.

ones: Similar to zeros, this function takes in the number of rows and
columns and returns a matrix filled with ones.

eye: Takes in the number of rows and returns the corresponding identity
matrix.