Add statistics module
The statistics module defines several functions used to compute several
metrics that describe a dataset.

mean: Computes the average value of a dataset.

median: Computes the middle value using two private methods
(_median_odd and _median_even).

quantile: Computes the pth-percentie value of a dataset.

mode: Returns a list of the most common value(s) in a dataset.

data_range: Computes the range of a dataset.

standardize: Rescales the dataset such that the mean is zero.

variance: Computes the squared deviation from the mean.

standard_deviation: Computes the average deviation from the mean.

interquartile_range: Computes the difference between the 75%-ile and the
25%-ile.

covariance: Computes the variance between two variables and their mean in
tandem.

correlation: Computes the amount that two variables are linearly related.